:recycle: move api-related code to parsing module
